move resolution logic into its own class
